Save and use another computer
