+++ server/controllers/add-question.js
@@ -10,7 +10,9 @@ const addQuestion = (req, res) => {
     meetup: req.params.id,
     title: req.body.title,
     body: req.body.question,
-    votes: 0,
+    upvotes: 0,
+    downvotes: 0,
+    votedBy: [],
   };
   questions.push(newQuestion);
   fs.writeFileSync(path.resolve(__dirname, '../data/questions.json'), JSON.stringify(questions, null, 2));
